# docs/api/openapi.yaml
openapi: 3.0.0
info:
  title: Describe It - Spanish Learning API
  version: 1.0.0
  description: |
    A comprehensive API for visual-based Spanish learning through AI-powered image descriptions, 
    Q&A generation, vocabulary extraction, and translation services.
    
    This API enables users to:
    - Generate multi-style image descriptions in Spanish and English
    - Create interactive Q&A sessions for comprehension testing
    - Extract and translate vocabulary from descriptions
    - Track learning progress and export data
    - Search and manage images for learning content
    
  contact:
    name: Describe It Support
    email: support@describe-it.app
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://describe-it.vercel.app/api
    description: Production server
  - url: http://localhost:3000/api
    description: Development server
security:
  - ApiKeyAuth: []
  - BearerAuth: []
paths:
...
  /progress/track:
    post:
      tags:
        - Progress
      summary: Track Learning Progress
      description: Record user learning progress and analytics
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProgressTrackingRequest'
      responses:
        '200':
          description: Progress tracked successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProgressResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ProgressTrackingRequest:
      type: object
      required:
        - sessionId
        - action
      properties:
        sessionId:
          type: string
          description: Unique session identifier
          example: "session_123456789"
        action:
          type: string
          enum: ["start", "description_generated", "qa_completed", "vocabulary_saved"]
          description: Action being tracked
          example: "qa_completed"
        metadata:
          type: object
          description: Additional metadata for the action
          example:
            correctAnswers: 4
            totalQuestions: 5
            timeSpent: 120
...
    ProgressResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        progress:
          $ref: '#/components/schemas/LearningProgress'
...
    LearningProgress:
      type: object
      properties:
        sessionId:
          type: string
          example: "session_123456789"
        totalDescriptions:
          type: integer
          example: 5
        totalQuestions:
          type: integer
          example: 25
        correctAnswers:
          type: integer
          example: 20
        vocabularyLearned:
          type: integer
          example: 45
        timeSpent:
          type: integer
          description: Time spent in seconds
          example: 1800
        accuracy:
          type: number
          format: float
          example: 0.80
        lastActivity:
          type: string
          format: date-time
          example: "2024-01-15T10:30:00Z"

    # Error Schemas
...